requirements.txt
flet>=0.22.0
coqui-tts
sounddevice
scipy
matplotlib
pytest
